power_supply: Prevent double-registration to hwmon
Power supplies can get registered with both the thermal and hwmon subsystems,
of which the second is optional. This, however, leads to a double-registration
to hwmon because the thermal subsystem registers with hwmon a second time.

The thermal zone API has a flag for suppressing hwmon registration, which is
what we now use to prevent this double-registration. Thermal values will still
be available through our own hwmon registration. If POWER_SUPPLY_HWMON is
disabled then we do not set the flag so that the power supply will be visible
to hwmon as a thermal device only.

This is a bug that may have existed for over a decade. It can be detected
using the 'sensors' utility from lm-sensors on any ACPI or SoC device that
uses the power supply core. Systems with the bug will display all of their
power supplies twice. The tool will run into errors like these on the second
instance:

ERROR: Can't get value of subfeature in0_input: Kernel interface error
ERROR: Can't get value of subfeature in0_min: Kernel interface error
ERROR: Can't get value of subfeature in0_max: Kernel interface error
ERROR: Can't get value of subfeature in0_average: Kernel interface error

These are caused by the duplicate hwmon object name and are fixed by this patch.